Begin bridging model maker assets to game assets
At the moment it's a bit silly to have separate asset types for the game, engine, and model maker. So construct the in-memory game assets directory from the model maker assets.
This will speed up development by preventing the constant need to export to various formats, that all pretty much have the same information anyway.